Move DataPermission types into metabase-types
