+++ enterprise/frontend/src/metabase-enterprise/advanced_permissions/reducer.ts
@@ -7,16 +7,17 @@ import {
   SAVE_DATA_PERMISSIONS,
   UPDATE_DATA_PERMISSION,
 } from "metabase/admin/permissions/permissions";
-import {
-  DataPermission,
-  DataPermissionValue,
-  type EntityId,
-} from "metabase/admin/permissions/types";
+import type { EntityId } from "metabase/admin/permissions/types";
 import {
   DATABASES_BASE_PATH,
   GROUPS_BASE_PATH,
 } from "metabase/admin/permissions/utils/urls";
-import type { GroupId, Impersonation } from "metabase-types/api";
+import {
+  DataPermission,
+  DataPermissionValue,
+  type GroupId,
+  type Impersonation,
+} from "metabase-types/api";
 
 export const getImpersonatedPostAction = (
   entityId: EntityId,
