+++ enterprise/frontend/src/metabase-enterprise/feature_level_permissions/permission-management/transforms-permission.ts
@@ -7,16 +7,19 @@ import {
   getPermissionWarningModal,
 } from "metabase/admin/permissions/selectors/confirmations";
 import {
-  DataPermission,
   DataPermissionType,
-  DataPermissionValue,
   type EntityId,
   type PermissionOption,
   type PermissionSectionConfig,
   type PermissionSubject,
 } from "metabase/admin/permissions/types";
 import { getSchemasPermission } from "metabase/admin/permissions/utils/graph";
-import type { Group, GroupsPermissions } from "metabase-types/api";
+import {
+  DataPermission,
+  DataPermissionValue,
+  type Group,
+  type GroupsPermissions,
+} from "metabase-types/api";
 
 export const TRANSFORMS_PERMISSION_OPTIONS: Record<string, PermissionOption> = {
   no: {
