+++ enterprise/frontend/src/metabase-enterprise/feature_level_permissions/permission-management/transforms-permission.unit.spec.ts
@@ -1,5 +1,8 @@
-import { DataPermissionValue } from "metabase/admin/permissions/types";
-import type { Group, GroupsPermissions } from "metabase-types/api";
+import {
+  DataPermissionValue,
+  type Group,
+  type GroupsPermissions,
+} from "metabase-types/api";
 
 import {
   TRANSFORMS_PERMISSION_OPTIONS,
