+++ frontend/src/metabase-types/api/permissions.ts
@@ -1,8 +1,4 @@
 import type {
-  DataPermission,
-  DataPermissionValue,
-} from "metabase/admin/permissions/types";
-import type {
   CollectionId,
   DatabaseId,
   SchemaName,
@@ -11,6 +7,40 @@ import type {
 
 import type { GroupId } from "./group";
 import type { UserAttributeKey } from "./user";
+
+export enum DataPermission {
+  VIEW_DATA = "view-data",
+  CREATE_QUERIES = "create-queries",
+  DOWNLOAD = "download",
+  DATA_MODEL = "data-model",
+  DETAILS = "details",
+  TRANSFORMS = "transforms",
+  COLLECTIONS = "collections",
+}
+
+export enum DataPermissionValue {
+  BLOCKED = "blocked",
+  CONTROLLED = "controlled",
+  IMPERSONATED = "impersonated",
+  LEGACY_NO_SELF_SERVICE = "legacy-no-self-service",
+  NO = "no",
+  QUERY_BUILDER = "query-builder",
+  QUERY_BUILDER_AND_NATIVE = "query-builder-and-native",
+  SANDBOXED = "sandboxed",
+  UNRESTRICTED = "unrestricted",
+  // download specific values
+  NONE = "none",
+  LIMITED = "limited",
+  FULL = "full",
+  // details specific values
+  YES = "yes",
+  // data model specific values
+  ALL = "all",
+  //collections
+  WRITE = "write",
+  READ = "read",
+  //NONE = "none", //shared with download above
+}
 
 export type PermissionsGraph = {
   groups: GroupsPermissions;
