+++ frontend/src/metabase/admin/permissions/components/PermissionsConfirm.tsx
@@ -5,9 +5,9 @@ import type {
   GroupPermissionsDiff,
   PermissionsGraphDiff,
 } from "metabase/admin/permissions/types";
-import { DataPermissionValue } from "metabase/admin/permissions/types";
 import CS from "metabase/css/core/index.css";
 import { Tooltip } from "metabase/ui";
+import { DataPermissionValue } from "metabase-types/api";
 
 const GroupName = ({ group }: { group: GroupPermissionsDiff }) => (
   <span className={CS.textBrand}>{group.name}</span>
