+++ frontend/src/metabase/admin/permissions/utils/graph/data-permissions/utils.ts
@@ -1,5 +1,5 @@
-import { DataPermissionValue } from "metabase/admin/permissions/types";
 import { PLUGIN_ADVANCED_PERMISSIONS } from "metabase/plugins";
+import { DataPermissionValue } from "metabase-types/api";
 
 export const isRestrictivePermission = (value: DataPermissionValue) =>
   value === DataPermissionValue.NO ||
